+++ classes/Dexterity.lua
@@ -36,6 +36,7 @@ function UT.Dexterity:setInfiniteStamina(value)
     _G.CloneClass(PlayerMovement)
     if value then
         function PlayerMovement:_change_stamina() end
+
         function PlayerMovement:is_stamina_drained() return false end
     else
         PlayerMovement._change_stamina = PlayerMovement.orig._change_stamina
@@ -129,6 +130,52 @@ function UT.Dexterity:setInstantReload(value)
     end
 end
 
+function UT.Dexterity:setShootThroughWalls(value)
+    _G.CloneClass(RaycastWeaponBase)
+    _G.CloneClass(NewRaycastWeaponBase)
+
+    local player = managers.player:player_unit()
+    if value ~= nil then
+        if not player or not alive(player) then
+            UT.Dexterity:setShootThroughWalls(value)
+            return
+        end
+    end
+
+    for _, selection in pairs(player:inventory()._available_selections) do
+        local unitBase = selection.unit:base()
+        if value then
+            RaycastWeaponBase._can_shoot_through_shield_old = RaycastWeaponBase._can_shoot_through_shield
+            RaycastWeaponBase._can_shoot_through_shield = true
+            RaycastWeaponBase._can_shoot_through_wall_old = RaycastWeaponBase._can_shoot_through_wall
+            RaycastWeaponBase._can_shoot_through_wall = true
+            NewRaycastWeaponBase._can_shoot_through_shield_old = NewRaycastWeaponBase._can_shoot_through_shield
+            NewRaycastWeaponBase._can_shoot_through_shield = true
+            NewRaycastWeaponBase._can_shoot_through_wall_old = NewRaycastWeaponBase._can_shoot_through_wall
+            NewRaycastWeaponBase._can_shoot_through_wall = true
+
+            unitBase._bullet_slotmask_old = unitBase._bullet_slotmask
+            unitBase._bullet_slotmask = World:make_slot_mask(7, 11, 12, 14, 16, 17, 18, 21, 22, 25, 26, 33, 34, 35)
+        else
+            RaycastWeaponBase._can_shoot_through_shield = RaycastWeaponBase._can_shoot_through_shield_old
+            RaycastWeaponBase._can_shoot_through_wall = RaycastWeaponBase._can_shoot_through_wall_old
+            NewRaycastWeaponBase._can_shoot_through_shield = NewRaycastWeaponBase._can_shoot_through_shield_old
+            NewRaycastWeaponBase._can_shoot_through_wall = NewRaycastWeaponBase._can_shoot_through_wall_old
+
+            if unitBase._bullet_slotmask_old then
+                unitBase._bullet_slotmask = unitBase._bullet_slotmask_old
+                unitBase._bullet_slotmask_old = nil
+            end
+        end
+    end
+
+    if value then
+        UT:addAlert("ut_alert_shoot_through_walls_enabled", UT.colors.success)
+    else
+        UT:addAlert("ut_alert_shoot_through_walls_disabled", UT.colors.success)
+    end
+end
+
 function UT.Dexterity:setNoRecoil(value)
     _G.CloneClass(NewRaycastWeaponBase)
     if value then
@@ -155,6 +202,7 @@ function UT.Dexterity:setUnlimitedAmmo(value)
             self:set_ammo_total(self:get_ammo_max())
             return self:get_ammo_remaining_in_clip() == 0
         end
+
         function SawWeaponBase:clip_empty()
             self:set_ammo_total(self:get_ammo_max())
             return self:get_ammo_remaining_in_clip() == 0
@@ -189,10 +237,15 @@ function UT.Dexterity:setUnlimitedEquipment(value)
     _G.CloneClass(PlayerManager)
     if value then
         function BaseInteractionExt:_has_required_upgrade() return true end
+
         function BaseInteractionExt:_has_required_deployable() return true end
+
         function BaseInteractionExt:can_interact() return true end
+
         function PlayerManager:on_used_body_bag() end
+
         function PlayerManager:remove_equipment() end
+
         function PlayerManager:remove_special() end
     else
         BaseInteractionExt._has_required_upgrade = BaseInteractionExt.orig._has_required_upgrade
@@ -244,6 +297,7 @@ function UT.Dexterity:setDamageMultiplier(value, multiplier)
             end
             return self.orig.damage_bullet(self, attack_data)
         end
+
         function CopDamage:damage_melee(attack_data)
             if attack_data.attacker_unit == managers.player:player_unit() then
                 attack_data.damage = multiplier * 10